Make swank command default to binding to localhost and add 2dn optional argument which is the hostname to bind to.
The swank command used to bind to '0.0.0.0' by default, i.e. was open
to the world which is a dangerous default.  The default has now been
changed to 'localhost' which is safer.  The command now takes a second
optional argument which is the hostname to bind to -- pass '0.0.0.0'
to get the old behaviour.